requirements.txt
fastapi
uvicorn[standard]
requests
google-adk